Revert "Create a type package to just hold types"
This reverts commit daa0899907343e5cf1208a20d21fb767ca9db0df.
