+++ glusterd2/xlator/global.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"runtime/debug"
 
+	"github.com/gluster/glusterd2/glusterd2/xlator/options"
+
 	log "github.com/sirupsen/logrus"
 )
 
@@ -28,5 +30,21 @@ func Load() (err error) {
 	}
 	Xlators = xls
 
+	// Also prepare the option.Options map
+	loadOptions()
+
 	return
+}
+
+// loadOptions loads all available options into the options.Options map,
+// indexed as <xlator.ID>.<option.Key> for all available option keys
+func loadOptions() {
+	for _, xl := range Xlators {
+		for _, opt := range xl.Options {
+			for _, k := range opt.Key {
+				k := xl.ID + "." + k
+				options.Options[k] = opt
+			}
+		}
+	}
 }
