+++ glusterd2/xlator/options/options.go
@@ -1,12 +1,11 @@
 package options
 
-import (
-	"github.com/gluster/glusterd2/pkg/types"
-)
+// OptionType is a type which represents the type of xlator option.
+type OptionType int
 
 // These are all the possile values for the type OptionType.
 const (
-	OptionTypeAny types.OptionType = iota
+	OptionTypeAny OptionType = iota
 	OptionTypeStr
 	OptionTypeInt
 	OptionTypeSizet
@@ -24,17 +23,24 @@ const (
 	OptionTypeClientAuthAddr
 )
 
+// OptionValidateType is a type which represents how the value of xlator
+// option should be validated.
+type OptionValidateType int
+
 // These are all the possile values for the type OptionValidateType
 const (
-	OptionValidateBoth types.OptionValidateType = iota
+	OptionValidateBoth OptionValidateType = iota
 	OptionValidateMin
 	OptionValidateMax
 )
 
-// These are the available types.OptionFlags
+// OptionFlag is the type representing the flags of an Option
+type OptionFlag uint
+
+// These are the available OptionFlags
 const (
-	OptionFlagNone     types.OptionFlag = 0
-	OptionFlagSettable                  = 1 << iota
+	OptionFlagNone     = 0
+	OptionFlagSettable = 1 << iota
 	OptionFlagClientOpt
 	OptionFlagGlobal
 	OptionFlagForce
@@ -44,9 +50,20 @@ const (
 
 // Option is a struct which represents one single xlator option exported by
 // the translator.
-// Embedding the actual type declared in package types, to allow custom methods.
 type Option struct {
-	*types.Option
+	Key          []string
+	Type         OptionType
+	Value        []string
+	DefaultValue string
+	Description  string
+	Min          float64
+	Max          float64
+	ValidateType OptionValidateType
+	OpVersion    []uint32
+	Deprecated   []uint32
+	Flags        uint32
+	Tags         []string
+	SetKey       string
 }
 
 // Validate checks if the given value string can be set as the value for the
