--- pkg/types/doc.go
@@ -1,4 +1,0 @@
-// Package types collects the different type declarations from across different
-// GD2 pacakges in a single package. This helps avoid circular dependencies in
-// some cases
-package types
