--- pkg/types/xlator.go
@@ -1,7 +1,0 @@
-package types
-
-// Xlator represents a GlusterFS xlator
-type Xlator struct {
-	ID      string
-	Options []*Option
-}
